rename options to something more specific
