+++ src/plugins/plot/barGraph/inspector/BarGraphOptions.vue
